// sentadel-fe-next-master/src/api/queries/key.ts
export enum MUTATION_KEY {
  ADD_NEW_PARTNER = "ADD_NEW_PARTNER",
  ADD_NEW_LOAN = "ADD_NEW_LOAN",
  CREATE_QUEUE_REQUEST = "CREATE_QUEUE_REQUEST",
  GET_PARTNERS = "GET_PARTNERS",
  ADD_NEW_REPAYMENT = "ADD_NEW_REPAYMENT",
  MANAGE_INVOICE_STATUS = "MANAGE_INVOICE_STATUS",
  UPDATE_GRADE_INFORMATION = "UPDATE_GRADE_INFORMATION",
  MANAGE_ADDRESS = "MANAGE_ADDRESS",
}

export enum QUERY_KEY {
  GET_ALL_COORDINATOR_LIST = "GET_ALL_COORDINATOR_LIST",
  GET_GROUPED_PARTNERS = "GET_GROUPED_PARTNERS",
  GET_PARTNERS = "GET_PARTNERS",
  GET_LOAN_LIST = "GET_LOAN_LIST",
  GET_GOODS_INFO = "GET_GOODS_INFO",
  GET_DELIVERY_WITH_STATUS_ACCUM = "GET_DELIVERY_WITH_STATUS_ACCUM",
  GET_PURCHASE_DELIVERY_DETAIL = "GET_PURCHASE_DELIVERY_DETAIL",
  GET_PURCHASE_INVOICE_DETAIL = "GET_PURCHASE_INVOICE_DETAIL",
  GET_INVOICE_LIST = "GET_INVOICE_LIST",
  GET_PENDING_VALIDATION = "GET_PENDING_VALIDATION",
  GET_STOCK_LIST = "GET_STOCK_LIST",
  GET_STOCK_DETAIL = "GET_STOCK_DETAIL",
  GET_ALL_GRADE = "GET_ALL_GRADE",
  GET_SALES_CODE = "GET_SALES_CODE",
  GET_GROUPING_LIST = "GET_GROUPING_LIST",
  GET_GROUPING_DETAIL = "GET_GROUPING_DETAIL",
}